+++ library/core/src/main/java/com/google/android/exoplayer2/source/ads/AdsLoader.java
@@ -16,6 +16,7 @@
 package com.google.android.exoplayer2.source.ads;
 
 import android.support.annotation.Nullable;
+import android.view.View;
 import android.view.ViewGroup;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.Player;
@@ -31,16 +32,16 @@ import java.io.IOException;
  * with a new copy of the current {@link AdPlaybackState} whenever further information about ads
  * becomes known (for example, when an ad media URI is available, or an ad has played to the end).
  *
- * <p>{@link #start(EventListener, ViewGroup)} will be called when the ads media source first
+ * <p>{@link #start(EventListener, AdViewProvider)} will be called when the ads media source first
  * initializes, at which point the loader can request ads. If the player enters the background,
  * {@link #stop()} will be called. Loaders should maintain any ad playback state in preparation for
- * a later call to {@link #start(EventListener, ViewGroup)}. If an ad is playing when the player is
- * detached, update the ad playback state with the current playback position using {@link
+ * a later call to {@link #start(EventListener, AdViewProvider)}. If an ad is playing when the
+ * player is detached, update the ad playback state with the current playback position using {@link
  * AdPlaybackState#withAdResumePositionUs(long)}.
  *
  * <p>If {@link EventListener#onAdPlaybackState(AdPlaybackState)} has been called, the
- * implementation of {@link #start(EventListener, ViewGroup)} should invoke the same listener to
- * provide the existing playback state to the new player.
+ * implementation of {@link #start(EventListener, AdViewProvider)} should invoke the same listener
+ * to provide the existing playback state to the new player.
  */
 public interface AdsLoader {
 
@@ -69,6 +70,25 @@ public interface AdsLoader {
     default void onAdTapped() {}
   }
 
+  /** Provides views for the ad UI. */
+  interface AdViewProvider {
+
+    /** Returns the {@link ViewGroup} on top of the player that will show any ad UI. */
+    ViewGroup getAdViewGroup();
+
+    /**
+     * Returns an array of views that are shown on top of the ad view group, but that are essential
+     * for controlling playback and should be excluded from ad viewability measurements by the
+     * {@link AdsLoader} (if it supports this).
+     *
+     * <p>Each view must be either a fully transparent overlay (for capturing touch events), or a
+     * small piece of transient UI that is essential to the user experience of playback (such as a
+     * button to pause/resume playback or a transient full-screen or cast button). For more
+     * information see the documentation for your ads loader.
+     */
+    View[] getAdOverlayViews();
+  }
+
   // Methods called by the application.
 
   /**
@@ -95,8 +115,8 @@ public interface AdsLoader {
 
   /**
    * Sets the supported content types for ad media. Must be called before the first call to {@link
-   * #start(EventListener, ViewGroup)}. Subsequent calls may be ignored. Called on the main thread
-   * by {@link AdsMediaSource}.
+   * #start(EventListener, AdViewProvider)}. Subsequent calls may be ignored. Called on the main
+   * thread by {@link AdsMediaSource}.
    *
    * @param contentTypes The supported content types for ad media. Each element must be one of
    *     {@link C#TYPE_DASH}, {@link C#TYPE_HLS}, {@link C#TYPE_SS} and {@link C#TYPE_OTHER}.
@@ -107,9 +127,9 @@ public interface AdsLoader {
    * Starts using the ads loader for playback. Called on the main thread by {@link AdsMediaSource}.
    *
    * @param eventListener Listener for ads loader events.
-   * @param adUiViewGroup A {@link ViewGroup} on top of the player that will show any ad UI.
+   * @param adViewProvider Provider of views for the ad UI.
    */
-  void start(EventListener eventListener, ViewGroup adUiViewGroup);
+  void start(EventListener eventListener, AdViewProvider adViewProvider);
 
   /**
    * Stops using the ads loader for playback and deregisters the event listener. Called on the main