+++ RELEASENOTES.md
@@ -50,6 +50,10 @@
   * Remove deprecated `AdsMediaSource` constructors. Listen for media source
     events using `AdsMediaSource.addEventListener`, and ad interaction events by
     adding a listener when building `ImaAdsLoader`.
+  * Allow apps to register playback-related obstructing views that are on top of
+    their ad display containers via `AdsLoader.AdViewProvider`. `PlayerView`
+    implements this interface and will register its control view. This makes it
+    possible for ad loading SDKs to calculate ad viewability accurately.
 * Remove `player` and `isTopLevelSource` parameters from `MediaSource.prepare`.
 * VP9 extension: Remove RGB output mode and libyuv dependency, and switch to
   surface YUV output as the default. Remove constructor parameters `scaleToFit`